Rename Imbens Manski funcs file.
