+++ src/thesis/simple_model/task_simple_model_sims.py
@@ -19,8 +19,8 @@ class _Arguments(NamedTuple):
     local_ates: LocalATEs
     pscore_hi: float = 0.6
     alpha: float = 0.05
-    n_boot: int = 1_000
-    n_sims: int = 1_000
+    n_boot: int = 2
+    n_sims: int = 2
     rng: np.random.Generator = RNG
